refactor(tagging): Misc refactor with tagging
- move logic out of the web module
- move tagging flag to resource type level
