+++ swabbie-core/src/main/kotlin/com/netflix/spinnaker/swabbie/tagging/ResourceEntityTagger.kt
@@ -1,0 +1,98 @@
+/*
+ * Copyright 2018 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.netflix.spinnaker.swabbie.tagging
+
+import com.netflix.spinnaker.config.SwabbieProperties
+import com.netflix.spinnaker.moniker.frigga.FriggaReflectiveNamer
+import com.netflix.spinnaker.swabbie.model.*
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+import org.springframework.stereotype.Component
+import java.time.Clock
+
+@Component
+class ResourceEntityTagger(
+  private val clock: Clock,
+  private val swabbieProperties: SwabbieProperties,
+  private val taggingService: TaggingService
+) : ResourceTagger {
+  private val log: Logger = LoggerFactory.getLogger(javaClass)
+  override fun tag(markedResource: MarkedResource, workConfiguration: WorkConfiguration, description: String) {
+    if (!workConfiguration.entityTaggingEnabled) {
+      log.debug("Skipping tagging of resource {}", markedResource)
+      return
+    }
+
+    log.info("tagging resource {}", markedResource)
+    taggingService.tag(
+      UpsertEntityTagsRequest(
+      entityRef = EntityRef(
+        entityType = markedResource.resourceType.toLowerCase(),
+        cloudProvider = markedResource.cloudProvider,
+        entityId = markedResource.resourceId,
+        region = workConfiguration.location,
+        account = workConfiguration.account.name!!
+      ),
+      tags = listOf(
+        EntityTag(
+          namespace = "swabbie:${workConfiguration.namespace.toLowerCase()}",
+          value = TagValue(message = tagMessage(markedResource))
+        )
+      ),
+      application = FriggaReflectiveNamer().deriveMoniker(markedResource).app ?: "swabbie",
+      description = description
+      )
+    )
+  }
+
+  private fun tagMessage(markedResource: MarkedResource): String {
+    return markedResource.summaries
+      .joinToString(", ") {
+        it.description
+      }.also { summary ->
+        val time = markedResource.humanReadableDeletionTime(clock)
+        "Scheduled to be cleaned up on $time<br /> \n " +
+          "* $summary <br /> \n" +
+          "* Click <a href='${swabbieProperties.optOutBaseUrl}' target='_blank'>here</a> to opt out."
+      }
+  }
+
+  override fun unTag(markedResource: MarkedResource, workConfiguration: WorkConfiguration, description: String) {
+    if (!workConfiguration.entityTaggingEnabled) {
+      log.debug("Skipping removing tag from resource {}", markedResource)
+      return
+    }
+
+    log.info("removing tagging resource {}", markedResource)
+    taggingService.removeTag(
+      DeleteEntityTagsRequest(
+        id = tagId(workConfiguration, markedResource),
+        application = FriggaReflectiveNamer().deriveMoniker(markedResource).app ?: "swabbie",
+        description = description
+      )
+    )
+  }
+
+  private fun tagId(workConfiguration: WorkConfiguration, markedResource: MarkedResource): String =
+    String.format("%s:%s:%s:%s:%s",
+      workConfiguration.cloudProvider,
+      workConfiguration.resourceType.toLowerCase(),
+      markedResource.resourceId,
+      workConfiguration.account.accountId,
+      workConfiguration.location
+    )
+}
