+++ swabbie-orca/src/main/kotlin/com/netflix/spinnaker/swabbie/orca/tagging/EntityTaggingService.kt
@@ -14,16 +14,16 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.swabbie.tagging
+package com.netflix.spinnaker.swabbie.orca.tagging
 
-import com.netflix.spinnaker.swabbie.TagRequest
-import com.netflix.spinnaker.swabbie.TaggingService
 import com.netflix.spinnaker.swabbie.orca.OrcaJob
 import com.netflix.spinnaker.swabbie.orca.OrcaService
 import com.netflix.spinnaker.swabbie.orca.OrchestrationRequest
-import org.springframework.stereotype.Component
+import com.netflix.spinnaker.swabbie.tagging.DeleteEntityTagsRequest
+import com.netflix.spinnaker.swabbie.tagging.TagRequest
+import com.netflix.spinnaker.swabbie.tagging.TaggingService
+import com.netflix.spinnaker.swabbie.tagging.UpsertEntityTagsRequest
 
-@Component
 class EntityTaggingService(
   private val orcaService: OrcaService
 ) : TaggingService {
