+++ swabbie-orca/swabbie-orca.gradle
@@ -15,6 +15,7 @@
  */
 
 dependencies {
+  compile project(":swabbie-core")
   compile project(":swabbie-retrofit")
   compile "com.netflix.spinnaker.moniker:moniker:+"
 }
